editoast: move unit tests from schema.railjson into models.infra
